Refatorar estrutura das pastas e dos componentes
