+++ packages/web/components/Modal/LoginModal.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Form, Actions, InputField } from '../Form';
-import Link from '../Link';
-import Button from '../Button';
+import { Link } from '../Link';
+import { Button } from '../Button';
 import { StyledLoginModal } from './styles';
 
 import { useModal, useAuth } from '../../hooks';
