+++ packages/web/components/Modal/index.js
@@ -1,1 +1,2 @@
-export * from './ModalProvider';
+export { default as ModalProvider } from './ModalProvider';
+export { default as LoginModal } from './LoginModal';
